refactor(issues): split the issues.go god-file by concern
Mirrors the detector-layer split: one 657-line file → cohesive single-purpose
files matching the rest of the package.
- compose.go: Provider/ComposeStats + Compose/ComposeWithStats (the pipeline).
- source_conditions.go: the generic CRD .status.conditions source + curated-kind
  exclusion + transient suppression.
- normalize.go: Detection→Issue (fromProblem/classifyIssue/resolveGroup).
- dedupe.go: evidence-level dedup (scheduling-over-problem, workload-over-child).
- filters.go: severity/kind + cluster-scoped-access filters + SeverityRank.
No logic change; pure reorganization to cut the high-conflict god-file.
